Separated out the db functionality into a different module
